+++ rn-vr.config.js
@@ -1,9 +1,9 @@
 module.exports = {
   devices: [
-    // {
-    //   platform: "android",
-    //   name: "Pixel_8_API_35",
-    // },
+    {
+      platform: "android",
+      name: "Pixel_8_API_35",
+    },
     {
       platform: "ios",
       name: "iPhone 15",
